My first customized keyboard and keymap settings
It's a PCB from kbdfans.com
(https://kbdfans.com/collections/60/products/dz60rgb-ansi-mechanical-keyboard-pcb)

build:
    make dztech/dz60rgb_ansi/v1:ckyoog

flash:
    make dztech/dz60rgb_ansi/v1:ckyoog:flash

"v1" is ARM, "v2" is AVR

Its DFU (bootloader) is STM32 BOOTLOADER. The following is the
screenshot from `dmesg'.
	[247129.278661] usb 1-1.2: Product: STM32  BOOTLOADER
	[247129.278661] usb 1-1.2: Manufacturer: STMicroelectronics
	[247129.278662] usb 1-1.2: SerialNumber: 206E31622037

The way to enter DFU mode is to unplug the keyboard, press and hold key
ESC, plug in the keyboard and wait a second before releasing the key.

My changes are based on master of https://github.com/qmk/qmk_firmware,
which is several commits ahead of the last tag 0.7.106 as of 2019/12/31.